requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=9.5.0

# Video/Audio processing
opencv-python>=4.7.0
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
scipy>=1.10.0
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Optional: GPU support
# faiss-gpu>=1.7.4
